+++ lib/stoplight/light.rb
@@ -2,9 +2,6 @@
 
 module Stoplight
   class Light
-    # @return [Integer]
-    DEFAULT_THRESHOLD = 3
-
     # @return [Array<Exception>]
     attr_reader :allowed_errors
 
@@ -23,7 +20,7 @@ module Stoplight
     end
 
     # @return [Object]
-    # @raise [Error::NoFallback]
+    # @raise [Error::RedLight]
     # @see #fallback
     # @see #green?
     def run
@@ -32,6 +29,11 @@ module Stoplight
       if green?
         run_code
       else
+        if Stoplight.attempts(name).zero?
+          message = "Switching #{name} stoplight from green to red."
+          Stoplight.notifiers.each { |notifier| notifier.notify(message) }
+        end
+
         run_fallback
       end
     end
@@ -55,17 +57,17 @@ module Stoplight
     # @param threshold [Integer]
     # @return [self]
     def with_threshold(threshold)
-      Stoplight.data_store.set_threshold(name, threshold.to_i)
+      Stoplight.set_threshold(name, threshold.to_i)
       self
     end
 
     # Attribute readers
 
     # @return [Object]
-    # @raise [Error::NoFallback]
+    # @raise [Error::RedLight]
     def fallback
       return @fallback if defined?(@fallback)
-      fail Error::NoFallback
+      fail Error::RedLight
     end
 
     # @return (see Stoplight.green?)
@@ -91,25 +93,25 @@ module Stoplight
 
     def run_code
       result = code.call
-      Stoplight.data_store.clear_failures(name)
+      Stoplight.clear_failures(name)
       result
     rescue => error
       if error_allowed?(error)
-        Stoplight.data_store.clear_failures(name)
+        Stoplight.clear_failures(name)
       else
-        Stoplight.data_store.record_failure(name, error)
+        Stoplight.record_failure(name, error)
       end
 
       raise
     end
 
     def run_fallback
-      Stoplight.data_store.record_attempt(name)
+      Stoplight.record_attempt(name)
       fallback.call
     end
 
     def sync_settings
-      Stoplight.data_store.set_threshold(name, threshold)
+      Stoplight.set_threshold(name, threshold)
     end
   end
 end