+++ lib/stoplight/notifier/base.rb
@@ -1,0 +1,11 @@
+# coding: utf-8
+
+module Stoplight
+  module Notifier
+    class Base
+      def notify(_message)
+        fail NotImplementedError
+      end
+    end
+  end
+end